requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.33
albumentationsx==2.0.11
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
lightning==2.5.5
lightning-utilities==0.15.2
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
segmentation_models_pytorch==0.5.0
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
stringzilla==4.0.14
sympy==1.14.0
tifffile==2025.9.30
timm==1.0.20
torch==2.8.0
torchinfo==1.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
